% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{area_units}
\alias{area_units}
\title{Area units}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 1911 rows and 3 columns.}
\usage{
data(area_units)
}
\description{
The area units spatial data set without any other census variables
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
