% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_census_columns}
\alias{clean_census_columns}
\title{Clean variables}
\usage{
clean_census_columns(long_df)
}
\description{
Cleans the variable names and splits them into columns in a long table.
}
\examples{
nz_dwelling_regions_long <- read_nz_census_data(dwelling_regions, include_gis = FALSE, long = TRUE, replace_confidential_values = NA_integer_)
cleaned_data <- clean_census_columns(nz_dwelling_regions_long)
ggplot(
  cleaned_data \%>\% filter(topic == "dwelling record type for occupied dwellings", variable == "Occupied Non-private Dwelling"),
  aes(x = year, y = value, colour = Description, group = Description)) +
  geom_line() +
  theme_bw()
}
