% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{dwelling_area_units}
\alias{dwelling_area_units}
\title{Dwelling area units data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1911 rows and 130 columns.}
\usage{
data(dwelling_area_units)
}
\description{
Dwelling data set at the area unit level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
