% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{dwelling_local_boards}
\alias{dwelling_local_boards}
\title{Dwelling local boards data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 130 columns.}
\usage{
data(dwelling_local_boards)
}
\description{
Dwelling data set at the local board level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
