% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{dwelling_meshblocks}
\alias{dwelling_meshblocks}
\title{Dwelling meshblocks data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 45989 rows and 130 columns.}
\usage{
data(dwelling_meshblocks)
}
\description{
Dwelling data set at the meshblock level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
