% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{dwelling_regions}
\alias{dwelling_regions}
\title{Dwelling regional data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 130 columns.}
\usage{
data(dwelling_regions)
}
\description{
Dwelling data set at the regional level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
