% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{dwelling_tas}
\alias{dwelling_tas}
\title{Dwelling territorial authority data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68 rows and 130 columns.}
\usage{
data(dwelling_tas)
}
\description{
Dwelling data set at the territorial authority level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
