% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{dwelling_wards}
\alias{dwelling_wards}
\title{Dwelling wards data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 258 rows and 130 columns.}
\usage{
data(dwelling_wards)
}
\description{
Dwelling data set at the ward level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
