% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_topics}
\alias{extract_topics}
\title{Extract topic}
\usage{
extract_topics(raw_topics, unique_topics = TRUE)
}
\arguments{
\item{raw_topics}{A vector of raw topics.}

\item{unique_topics}{Whether the results should be unique or not. Defaults to TRUE.}
}
\value{
A vector of topics
}
\description{
Extracts the topics for a raw census column from a vector strings.
}
