% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variable}
\usage{
extract_variables(raw_variables)
}
\arguments{
\item{raw_variables}{A vector of census columns containing the raw variables.}
}
\value{
A vector of variables
}
\description{
Extracts the variables for a raw census column from a vector strings.
}
