% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{family_area_units}
\alias{family_area_units}
\title{Family area units data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1911 rows and 100 columns.}
\usage{
data(family_area_units)
}
\description{
Family data set at the area unit level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
