% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{family_meshblocks}
\alias{family_meshblocks}
\title{Family meshblocks data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 45989 rows and 100 columns.}
\usage{
data(family_meshblocks)
}
\description{
Family data set at the meshblock level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
