% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{family_tas}
\alias{family_tas}
\title{Family territorial authority data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68 rows and 100 columns.}
\usage{
data(family_tas)
}
\description{
Family data set at the territorial authority level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
