% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{family_wards}
\alias{family_wards}
\title{Family wards data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 258 rows and 100 columns.}
\usage{
data(family_wards)
}
\description{
Family data set at the wards level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
