% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_area}
\alias{filter_by_area}
\title{Filters census data by area}
\usage{
filter_by_area(.data, geographic_level, geographic_filter, pattern)
}
\arguments{
\item{.data}{The data set to be filtered. Must contain a column with census spatial
names in it e.g. (Area_Code_and_Description).}

\item{geographic_level}{The geographic level of the input spatial data e.g. area_units.}

\item{geographic_filter}{The geographic level of the spatial filter e.g. region.}

\item{pattern}{The pattern to match the area to filter by e.g. 'Wellington Region'.}
}
\value{
A sf dataframe.
}
\description{
Filters a census table by another area e.g. meshblocks by territorial area or region.
It can only work if the geographic level is at a greater scale than the input data e.g.
meshblocks < area_units < wards < local boards < territoral authorities < regions
}
