% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{household_area_units}
\alias{household_area_units}
\title{Household area units data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1911 rows and 259 columns.}
\usage{
data(household_area_units)
}
\description{
Household data set at the area unit level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
