% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{household_local_boards}
\alias{household_local_boards}
\title{Household local boards data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 259 columns.}
\usage{
data(household_local_boards)
}
\description{
Household data set at the local board level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
