% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{household_meshblocks}
\alias{household_meshblocks}
\title{Household meshblocks data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 45989 rows and 259 columns.}
\usage{
data(household_meshblocks)
}
\description{
Household data set at the meshblock level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
