% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{household_regions}
\alias{household_regions}
\title{Household regional data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 259 columns.}
\usage{
data(household_regions)
}
\description{
Household data set at the regional level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
