% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{household_tas}
\alias{household_tas}
\title{Household territorial authority data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68 rows and 259 columns.}
\usage{
data(household_tas)
}
\description{
Household data set at the territorial authority level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
