% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{household_wards}
\alias{household_wards}
\title{Household wards data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 258 rows and 259 columns.}
\usage{
data(household_wards)
}
\description{
Household data set at the wards level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
