% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{individual_part_1_tas}
\alias{individual_part_1_tas}
\title{Individual part 1 territorial authority data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68 rows and 228 columns.}
\usage{
data(individual_part_1_tas)
}
\description{
Individual part 1 data set at the territorial authority level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
