% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{individual_part_3a_area_units}
\alias{individual_part_3a_area_units}
\title{Individual part 3a area units data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1911 rows and 358 columns.}
\usage{
data(individual_part_3a_area_units)
}
\description{
Individual part 3a data set at the area unit level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
