% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{individual_part_3a_regions}
\alias{individual_part_3a_regions}
\title{Individual part 3a regional data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 358 columns.}
\usage{
data(individual_part_3a_regions)
}
\description{
Individual part 3a data set at the regional level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
