% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{individual_part_3a_tas}
\alias{individual_part_3a_tas}
\title{Individual part 3a territorial authority data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68 rows and 358 columns.}
\usage{
data(individual_part_3a_tas)
}
\description{
Individual part 3a data set at the territorial authority level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
