% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{individual_part_3b_local_boards}
\alias{individual_part_3b_local_boards}
\title{Individual part 3b local boards data}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 138 columns.}
\usage{
data(individual_part_3b_local_boards)
}
\description{
Individual part 3b data set at the local board level
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
