% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{local_boards}
\alias{local_boards}
\title{Local boards}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 195 rows and 3 columns.}
\usage{
data(local_boards)
}
\description{
The local boards spatial data set without any other census variables
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
