% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{meshblocks}
\alias{meshblocks}
\title{Meshblocks}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 45989 rows and 29 columns.}
\usage{
data(meshblocks)
}
\description{
The meshblocks spatial data set without any other census variables
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
