% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_census_data}
\alias{read_census_data}
\title{Read in NZ census data}
\usage{
read_census_data(data_set, include_gis = TRUE, long = FALSE, crs = 2198,
  sp = FALSE, replace_confidential_values = NULL)
}
\arguments{
\item{data_set}{The data set to be passed. See what_data_sets.}

\item{include_gis}{If the data set should include the gis column (defaults to TRUE).}

\item{long}{If the data set should be in long or short form (defaults to wide).}

\item{crs}{The desired Coordinate Reference System of the data set (defaults to 2198). Only}

\item{sp}{Convert the simple features data frame to a sp object.}

\item{replace_confidential_values}{Replacement of confidential values ("..C") with another
value e.g. NA or 0 or anything for that matter!
important if the data includes a geometry column.}
}
\value{
It returns either a tibble or a simple features dataframe.
}
\description{
This function is a helper to read in the census data. Mainly meant for easy transformations
such as whether to include the GIS data, what its CRS system should be (defaults to 2198) and whether it
should come in the wide (default) or long format.
}
