% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_nz_census_data}
\alias{read_nz_census_data}
\title{Read in NZ census data}
\usage{
read_nz_census_data(data_set, replace_confidential_values = NULL,
  include_gis = TRUE, crs = 2198, long = FALSE)
}
\arguments{
\item{data_set}{The data set to be passed. See what_data_sets.}

\item{replace_confidential_values}{Replacement of confidential values ("..C") with another
value e.g. NA or 0 or anything for that matter!}

\item{include_gis}{If the data set should include the gis column (defaults to TRUE).}

\item{crs}{The desired Coordinate Reference System of the data set (defaults to 2198). Only
important if the data includes a geometry column.}

\item{long}{Whether the data should be returned in the long format or not.}
}
\value{
It returns either a tibble or a simple features dataframe.
}
\description{
This function is a helper to read in the census data. Mainly meant for easy transformations
such as whether to include the GIS data, what its CRS system should be (defaults to 2198) and whether it
should come in the wide (default) or long format.
}
