% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 17 rows and 3 columns.}
\usage{
data(regions)
}
\description{
The regions spatial data set without any other census variables
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
