% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_confidential}
\alias{replace_confidential}
\title{Replace confidential values}
\usage{
replace_confidential(.data, replacement_value = NA_integer_)
}
\arguments{
\item{.data}{The data set to have its confidential values removed.}

\item{replacement_value}{The value to replace the confidential ones with. Defaults to NA_integer.}
}
\description{
The NZ census commonly uses the notation of '..C' when values are below a
certain threshold that they may reveal private details of certain individuals.
However, it is often required in an analysis to replace these values and
convert the column to an integer (to include ..C they need to be character).
}
\details{
This function takes the data set, and a replacement value and replaces all of the ..C values.
Mainly for use within the transform_census function.
}
