% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_census_topics}
\alias{select_census_topics}
\title{Filter the census topics in a census dataframe}
\usage{
select_census_topics(.data, topics, exclude = FALSE)
}
\arguments{
\item{.data}{The data frame from which the topics should be selected/excluded from.}

\item{topics}{The census topics to be selected or excluded.}

\item{exclude}{Whether the topics should be selected or excluded.}
}
\value{
A dataframe
}
\description{
This is a helper function for quickly and easily selecting or excluding topics in a nz census data set.
}
