% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_topics_variables}
\alias{split_topics_variables}
\title{Split the topics and variables}
\usage{
split_topics_variables(topic_variable)
}
\description{
Splits the variavble column of the long data into years, topics and variables.
Only intended for use inside of clean_census_variables()
}
