% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_to_long}
\alias{table_to_long}
\title{Convert wide to long}
\usage{
table_to_long(data)
}
\arguments{
\item{data}{The data set to be transformed.}
}
\description{
Merely converts a census dataframe from wide to long.
}
