% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{tas}
\alias{tas}
\title{Territorial Authorities}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 68 rows and 3 columns.}
\usage{
data(tas)
}
\description{
The territorial authorities spatial data set without any other census variables
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
