% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzcensr.R
\docType{data}
\name{wards}
\alias{wards}
\title{Wards}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 258 rows and 3 columns.}
\usage{
data(wards)
}
\description{
The wards spatial data set without any other census variables
}
\author{
Phil Donovan \email{philiprogerdonovan@gmail.com}
}
\keyword{data}
