% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{what_data_sets}
\alias{what_data_sets}
\title{Get a table of census data available}
\usage{
what_data_sets(table_name = FALSE)
}
\arguments{
\item{table_name}{The name of the table that for a description. Defaults to false for all tables/}
}
\description{
A function which shows what data sets are available and what they're called. For an
indepth query of a data set specify the table.
}
