% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_census_data_functions.R
\name{nz_census_tables}
\alias{nz_census_tables}
\title{Get a table of census data available}
\usage{
nz_census_tables(table_name = data.frame())
}
\arguments{
\item{table_name}{The name of the table that for a description. Defaults to false for all tables/}
}
\value{
A table describing the data sets or data set.
}
\description{
A function which shows what data sets are available and what they're called. For an
indepth query of a data set specify the table.
}
